Added HTTP Request Class and moved RunCucumberTest Class to the correct path
