+++ ca-functional/src/test/java/uk/sky/ovp/ca/functional/models/HttpRequest.java
@@ -1,0 +1,25 @@
+package uk.sky.ovp.ca.functional.models;
+
+import io.restassured.response.Response;
+import org.springframework.stereotype.Component;
+
+@Component
+public class HttpRequest {
+    private Response response;
+
+    public Response getResponse() {
+        return response;
+    }
+
+    public void setResponse(Response response) {
+        this.response = response;
+    }
+
+    public String getResponseBody() {
+        return this.response.getBody().asString();
+    }
+
+    public int getResponseStatusCode() {
+        return this.response.getStatusCode();
+    }
+}
