+++ ca-functional/src/test/java/uk/sky/ovp/ca/functional/steps/PrivateEndpointSteps.java
@@ -3,6 +3,5 @@ package uk.sky.ovp.ca.functional.steps;
 import org.springframework.beans.factory.annotation.Autowired;
 
 public class PrivateEndpointSteps {
-    @Autowired
-    private PrivateController privateController;
+
 }
